PIA: Pet Adoption
When the program is run, a menu of actions will be displayed that will look like this:
Welcome to the pet adoption system, what would you like to do?
1. See the available dogs
2. See the description of a dog
3. Register a new member
4. Adopt your partner

The program Will do the next tasks:

Dog breeds: You can see the breeds of dogs that are available to adopt, through a list that includes the name of the dog follow by its breed, that will be inside a .txt file.

Pet description and care: You can see the description of each animal which helps to have a much more technical information that helps the adoption process and that can have all the needs covered by the family that plans to adopt. This is going to be in a second .txt file.

Register a dog: To add a new dog to the list we need to include its breed, if is female or male, approximate age (human age), weight (kg) and if any special care is needed like is there present a disease, what drugs needs to consume, kind of dog’s food, portions, etc. You must also write a name for the future member of a family to complete the register.

Adopt a dog: When someone wants to adopt a dog, when you want to check out, the program will ask you if you are sure and can take the responsibility of take care of the selected dog and to finish the process you will be asked for the name of the dog to give a message to say goodbye.